requirements.txt
moviepy==2.0.0
openai-whisper==20231117
requests==2.31.0
Pillow==10.1.0
edge-tts==6.1.9
google-generativeai==0.3.2